Add base datagrid component + basic first applications grid
